+++ lib/Skeniraj.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class Skeniraj extends StatefulWidget {
+    const Skeniraj({
+    Key key,
+  }) : super(key: key);
+  @override
+  _SkenirajState createState() => _SkenirajState();
+}
+
+class _SkenirajState extends State<Skeniraj> {
+  @override
+  Widget build(BuildContext context) {
+    return Center(child: Text("SKENIRAJ"),);
+  }
+}
